Adds --distribution flag, refactor k8s generation
This commit adds the --distribution flag as well as refactors the
generate.go file to one single file, kubernetes.go.

This allows for all future changes to simply be made within
kubernetes.go rather than two separate files with similar code
(generate.go + kubernetes.go)
