+++ cmd/apply.go
@@ -43,7 +43,7 @@ var applyCmd = &cobra.Command{
 			kubectlCommand = append(kubectlCommand, "--namespace", Namespace)
 		}
 
-		if err := pkgcmd.ExecuteKubectl(InputFiles, kubectlCommand...); err != nil {
+		if err := pkgcmd.CreateKubernetesArtifacts(InputFiles, false, kubectlCommand...); err != nil {
 			fmt.Println(err)
 			os.Exit(-1)
 		}
