+++ pkg/cmd/kubernetes_test.go
@@ -1,0 +1,46 @@
+/*
+Copyright 2017 The Kedge Authors All rights reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package cmd
+
+import (
+	"io/ioutil"
+	"os"
+	"testing"
+)
+
+func TestWriteObject(t *testing.T) {
+
+	err := writeObject([]byte("foobar"))
+	if err != nil {
+		t.Fatalf("Error FindAbsPath: %s", err)
+	}
+
+}
+
+func TestFindAbsPath(t *testing.T) {
+	tmpDir, err := ioutil.TempDir("", "absPathTest")
+	if err != nil {
+		t.Fatal("creating temp dir:", err)
+	}
+	defer os.RemoveAll(tmpDir)
+
+	path := findAbsPath("", tmpDir)
+
+	if path != tmpDir {
+		t.Fatal("Error getting correct absolute path")
+	}
+}
